Python API/requirements.txt
# Core Framework
fastapi==0.110.0
uvicorn==0.30.0
python-multipart==0.0.9

# LangChain
langchain==0.2.14
langchain-community==0.2.7
langchain-qdrant==0.1.1
langchain-text-splitters==0.2.4
openai==0.27.0
tiktoken==0.5.2

# Vector Database
qdrant-client==1.12.0

# Document Processing
pypdf==5.1.0
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.3
lxml==5.1.0
markdown==3.5.2
docx2txt==0.8

# Utilities
loguru==0.7.2
python-dotenv==1.0.0
pydantic-settings==2.3.0
